requirements.txt
streamlit
deepface
tf-keras
opencv-python-headless
pillow
numpy